Factor out peephole to own test file
[ghstack-poisoned]
